# config/openapi.yaml
...
openapi: 3.0.0
info:
  title: Portal API
  description: Portal REST API.
  version: 0.1.0
servers:
  - url: 'http://tnp-iiitr.herokuapp.com'
security:
  - bearerAuth: []
...
paths:
...
  /metadata/company:
    post:
      tags:
        - Metadata > Company
      summary: Create Company Profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                name: GOOGLE
                address: This is Test address
                phone: '8888888888'
                website: www.example.com
                category: IT
                sector: IT
                mainPOCName: Mahajan
                mainPOCPhone: '8888888888'
                mainPOCEmail: maha@example.com
                mainPOCPosition: HR
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: string
              example: null
        '405':
          description: Method Not Allowed
          content:
            application/json:
              schema:
                type: object
              examples:
                example-0:
                  summary: Email not verified
                  value:
                    msg: Method Not Allowed
                example-1:
                  summary: Recreating profile
                  value:
                    msg: Method Not Allowed
    get:
      tags:
        - Metadata > Company
      summary: Get Company Data
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
              example:
                name: Microsoft
                address: This is Test address
                phone: '8888888888'
                website: www.example.com
                category: IT
                sector: IT
                mainPOCName: Mahajan
                mainPOCPhone: '8888888888'
                mainPOCEmail: maha@example.com
                mainPOCPosition: HR
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              example:
                msg: Unauthorized access
    put:
      tags:
        - Metadata > Company
      summary: Update Company Profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                change:
                  name: Microsoft
      responses:
        '204':
          description: No Content
          content:
            text/plain:
              schema:
                type: string
              example: null
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                type: object
              example:
                msg: Unauthorized access
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...